backend/agents/langchain_agent/requirements.txt
# Base dependencies from original files
langchain-core>=0.1.21
langchain>=0.1.4
pydantic>=2.5.2
pydantic-core>=2.14.5

# FastAPI and server dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0

# For OpenAI LLM
langchain-openai>=0.0.2
openai>=1.3.0

# Additional utilities
requests>=2.31.0
aiohttp>=3.9.1
httpx>=0.25.2

# JSON handling
orjson>=3.9.10

# For typing in Python 3.9
typing_extensions>=4.8.0